+++ org.j3d.aviatrix3d-core/src/main/java/org/j3d/aviatrix3d/output/graphics/BaseSurface.java
@@ -18,9 +18,8 @@ import javax.media.opengl.*;
 import java.util.HashMap;
 import java.util.Map;
 
-import javax.vecmath.Matrix4f;
-import javax.vecmath.Point3f;
-
+import org.j3d.maths.vector.Matrix4d;
+import org.j3d.maths.vector.Point3d;
 import org.j3d.util.DefaultErrorReporter;
 import org.j3d.util.ErrorReporter;
 import org.j3d.util.I18nManager;
@@ -505,7 +504,7 @@ public abstract class BaseSurface
                                    int y,
                                    int layer,
                                    int subLayer,
-                                   Matrix4f matrix,
+                                   Matrix4d matrix,
                                    String deviceId,
                                    boolean useLastFound)
     {
@@ -539,7 +538,7 @@ public abstract class BaseSurface
                                           int y,
                                           int layer,
                                           int subLayer,
-                                          Point3f position,
+                                          Point3d position,
                                           String deviceId,
                                           boolean useLastFound)
     {
@@ -573,7 +572,7 @@ public abstract class BaseSurface
                                          int y,
                                          int layer,
                                          int subLayer,
-                                         Point3f position,
+                                         Point3d position,
                                          String deviceId,
                                          boolean useLastFound)
     {
