+++ org.j3d.aviatrix3d-core/src/main/java/org/j3d/aviatrix3d/picking/TransformPickTarget.java
@@ -13,7 +13,7 @@
 package org.j3d.aviatrix3d.picking;
 
 // External imports
-import javax.vecmath.Matrix4f;
+import org.j3d.maths.vector.Matrix4d;
 
 // Local imports
 // None
@@ -50,12 +50,12 @@ public interface TransformPickTarget extends PickTarget
      *
      * @param mat The matrix to copy the transform data to
      */
-    public void getTransform(Matrix4f mat);
+    public void getTransform(Matrix4d mat);
 
     /**
      * Get the inverse version of the local transform.
      *
      * @param mat The matrix to copy the transform data to
      */
-    public void getInverseTransform(Matrix4f mat);
+    public void getInverseTransform(Matrix4d mat);
 }
