+++ org.j3d.aviatrix3d-core/src/main/java/org/j3d/aviatrix3d/rendering/CullInstructions.java
@@ -13,7 +13,7 @@
 package org.j3d.aviatrix3d.rendering;
 
 // External imports
-import javax.vecmath.Matrix4f;
+import org.j3d.maths.vector.Matrix4d;
 
 // Local imports
 // None
@@ -53,7 +53,7 @@ public class CullInstructions
      * traversal process. If no transformation is needed. then set
      * {@link #hasTransform} to false and ignore this matrix.
      */
-    public Matrix4f localTransform;
+    public Matrix4d localTransform;
 
     /**
      * Initialise a new instance of this instruction.
@@ -61,7 +61,7 @@ public class CullInstructions
     public CullInstructions()
     {
         children = new Cullable[8];
-        localTransform = new Matrix4f();
+        localTransform = new Matrix4d();
         hasTransform = false;
     }
